spacing for leaderboard to be bigger
