+++ src/pages/TopMinersPage.tsx
@@ -1,9 +1,12 @@
 import React, { useMemo } from "react";
-import { Box } from "@mui/material";
+import { useMediaQuery, Box } from "@mui/material";
 import { useNavigate } from "react-router-dom";
 import { Page } from "../components/layout";
 import { TopMinersTable, SEO } from "../components";
 import { useAllMiners } from "../api";
+import theme from "../theme";
+
+import { LeaderboardSidebar } from "../components/leaderboard/LeaderboardSidebar";
 
 const TopMinersPage: React.FC = () => {
   const navigate = useNavigate();
@@ -45,6 +48,16 @@ const TopMinersPage: React.FC = () => {
     return [...minerStats].sort((a, b) => b.totalScore - a.totalScore);
   }, [minerStats]);
 
+  // Dashboard-like responsive logic
+  const isMobile = useMediaQuery(theme.breakpoints.down("sm"));
+  const isTablet = useMediaQuery(theme.breakpoints.between("sm", "md"));
+  const isLargeScreen = useMediaQuery(theme.breakpoints.up("xl"));
+  const showSidebarRight = useMediaQuery(theme.breakpoints.up("xl"));
+
+  // Dynamic sidebar width based on screen size (matching DashboardPage)
+  const sidebarWidth =
+    isMobile || isTablet ? "100%" : isLargeScreen ? "340px" : "300px";
+
   return (
     <Page title="Miner Leaderboard">
       <SEO
@@ -53,18 +66,65 @@ const TopMinersPage: React.FC = () => {
       />
       <Box
         sx={{
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          minHeight: { xs: "auto", md: "calc(100vh - 80px)" },
           width: "100%",
-          py: { xs: 2, sm: 0 },
+          height: showSidebarRight ? "calc(100vh - 64px)" : "auto",
+          display: "flex",
+          flexDirection: showSidebarRight ? "row" : "column",
+          gap: { xs: 2, sm: 2, md: 2.5, lg: 3 },
+          py: { xs: 2, sm: 2, md: 2.5, lg: 3 },
+          px: { xs: 2, sm: 2, md: 2.5, lg: 3 },
+          overflow: "hidden",
         }}
       >
-        <Box sx={{ maxWidth: 1200, width: "100%" }}>
-          <TopMinersTable
-            miners={sortedMinerStats}
-            isLoading={isLoadingMinerStats}
+        {/* Main Content Area */}
+        <Box
+          sx={{
+            flex: 1,
+            display: "flex",
+            flexDirection: "column",
+            gap: { xs: 2, sm: 1.5 },
+            minHeight: 0,
+            overflow: showSidebarRight ? "auto" : "visible",
+            minWidth: 0,
+            pr: showSidebarRight ? 1 : 0,
+            "&::-webkit-scrollbar": {
+              width: "8px",
+            },
+            "&::-webkit-scrollbar-track": {
+              backgroundColor: "transparent",
+            },
+            "&::-webkit-scrollbar-thumb": {
+              backgroundColor: "rgba(255, 255, 255, 0.1)",
+              borderRadius: "4px",
+              "&:hover": {
+                backgroundColor: "rgba(255, 255, 255, 0.2)",
+              },
+            },
+          }}
+        >
+          <Box sx={{ width: "100%" }}>
+            <TopMinersTable
+              miners={sortedMinerStats}
+              isLoading={isLoadingMinerStats}
+              onSelectMiner={handleSelectMiner}
+            />
+          </Box>
+        </Box>
+
+        {/* Right Sidebar - Spacer to match Dashboard Live Activity */}
+        <Box
+          sx={{
+            width: showSidebarRight ? sidebarWidth : "100%",
+            height: showSidebarRight ? "100%" : "auto",
+            maxHeight: showSidebarRight ? "100%" : "700px",
+            flexShrink: 0,
+            display: showSidebarRight ? "flex" : "none", // Hide on mobile if empty, or keep structure
+            flexDirection: "column",
+          }}
+        >
+          {/* Render extracted Sidebar Content here */}
+          <LeaderboardSidebar
+            miners={minerStats}
             onSelectMiner={handleSelectMiner}
           />
         </Box>
